init: move path operations to pkg/upath
